remove rename repo
